+++ modules/gis/src/index.d.ts
@@ -4,4 +4,4 @@ export {BinaryGeometryData, BinaryAttribute} from './types';
 // Functions
 export {geojsonToBinary} from './lib/geojson-to-binary';
 export {binaryToGeoJson} from './lib/binary-to-geojson';
-export {transformBinaryCoords, transformGeoJsonCoords} from './lib/reproject';
+export {transformBinaryCoords, transformGeoJsonCoords} from './lib/transform';
