+++ modules/gis/src/index.js
@@ -1,3 +1,3 @@
 export {geojsonToBinary} from './lib/geojson-to-binary';
 export {binaryToGeoJson} from './lib/binary-to-geojson';
-export {transformBinaryCoords, transformGeoJsonCoords} from './lib/reproject';
+export {transformBinaryCoords, transformGeoJsonCoords} from './lib/transform';
