+++ modules/gis/src/lib/transform.d.ts
