+++ modules/gis/test/index.js
@@ -1,3 +1,3 @@
 import './binary-to-geojson.spec';
 import './geojson-to-binary.spec';
-import './reproject.spec';
+import './transform.spec';
